vfs: Update atime/mtime/ctime
I noticed we weren't tracking these in the VFS when trying to see if
ssh+scp was working.  Even though that code is old and destined for the
dumpster, I'd rather fix it a little and not waste time wondering if I
actually changed a file or not.

This is far from perfect, especially in that we don't tell the
filesystem that we updated an inode, but that goes for all of the VFS.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>