Add support for attempting returns from panic
If you ever wanted to keep going after a panic, it just got a little
easier.  You can exit the monitor ('exit' or 'e'), like always.  This
commit cleans up a little so the core doesn't keep panicking.

This will rarely work well, since the system is toasted, but I've been able
to get back to userspace with it.  You might lose the core that panicked,
(e.g. MCP triggers a racy panic), but at least that core won't keep polling
the console in the monitor.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>