pci: Add a struct device field
bnx2x has a lot of other places that are still using pcidev->dev, but it's
passing them to dont-care macros.  It's hokey, and spatch doesn't seem to
want to catch them all.  It's also a little hard to spatch after doing the
native conversion; some places have akaros_pcidev->dev, meaning the 'd' of
TBDF.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>