ether: Fix ETHERMAXTU calculations
In commit f5e7ebb31ce7 ("Changes devether's Max Trans Unit to 1500")
I changed ETHERMAXTU to not account for the header size.  I changed places
that had maxmtu, but not places that directly used ETHERMAXMTU.

As a fun side note, I noticed the problem in devether, near the vlan case.
I was tempted to change it in a previous commit, but decided to do a
separate commit.  In doing so, I did a more thorough job and grepped for
ETHERMAXMTU and found the bugs in 8139 and 8169.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>