Update Linux compatibility shims/coccis
These were updates I made for the r8169.

There's also a potential bugfix in there for mlx4.  wait_for_completion()
should sleep, not busywait.  This pops up later in the patchset when we
make netif_carrier_ok() not just return true.  Incidentally, if we're going
to busywait, we should be using udelay.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>