net: tcp: Refactor the tcp header structs
v4 and v6 have the same TCP header (since it's the same protocol).  Plan 9
embeds them in TCP-IPv4 and TCP-IPv6 structs.  It behooves us to have a
common struct.  It could be used to clean up a bunch of tcp.c code, but for
now I just need it for a header helper.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>