net: tcp: Fix TSO for incoming connections
TSO was only being used on outbound connections.  The issue was that the
'flags' in tcpmtu for inbound connections was seg.flags, not tcb->flags.
So that bit we were setting was discarded.

The fix is to do the lookup (which requires the ifc) once we finally have
the tcb, which is after the conversation comes out of limbo.  Instead of
looking up the ifc twice, we can just store the uncounted ref.  We'll
probably have other uses for the ifc.

Note that the ifc->feat is not exactly the same as ether->netif->feat.  The
device's features are copied over to the IP interface during etherbind.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>