net: Add an option for RX checksum offload
Plan 9 drivers assume that this feature is always on, if available, and it
is unqueryable.

Linux drivers have a flag for this, and by setting that flag to zero, NICs
(e.g. r8169) turn off their feature since they think the user doesn't want
it.

It's likely this will turn on RX checksumming on bnx2x and mlx4.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>