mlx4: Use netif_stats for ifstat
Linux drivers have a way of exporting their stats to a struct netif_stats.
We have a way to parse netif_stats when we cat ether/ifstats.  Put those
together and you have a simple way to get stats from any Linux driver.

For mlx4, I had to turn on a few bits of code that we had commented out.
There were minor issues with the MAC address changing, but I think we can
ignore that for now.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>