Allow sys_halt_core() to monitor notif_pending
This lets the 2LS halt the core and wait for someone to touch
notif_pending, which is touched for any event delivery to the vcore.

2LSs that want to wake vcores without sending a full event will need to use
wake_vcore().  The force_ipi is to help them break out of spinning
uthreads.  We'll see if that's needed or not.  Some of this ties into the
long range 2LS changes I've been planning.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>