vmm: Halt vcore 0 instead of spinning
Spinning limits the sibling core's performance and limits the turbo mode
potential of the rest of the machine.

Instead, we'll mwait on notif_pending or an IPI.  Our ev_qs are already set
up to send IPIs with the notification, so things like syscalls and timers
will work fine.  If the 2LS enqueues a task thread - which are handled by
VC 0 - it'll make sure the vcore wakes up.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>