vmm: Provide a fast-path for IPIs in the kernel
Handling the IPI in userspace takes a couple usec.  If the target guest
pcore is loaded and running, such as when in greedy mode or under load, the
kernel can inject the IPI.

This helped perf bench pipe sched on tinycore in greedy mode, where the
guest was sending IPIs every 8-10 usec.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>