x86: Fix topology detection
The main thing is that we were off by one - the is_unique check would
compare us to ourselves, which will always fail.

The end result was that the number of numa nodes was always 0, which turns
off all of the other topology information.  I'm not sure how this ever
worked.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>