9ns: Add high-level support for symlinks
This creates links with SYS_symlink and extracts the info with
SYS_readlink.  You need a 9ns device that supports symlinks, such as the
yet-unmerged KFS.  Walking through symlinks doesn't work yet.

Similar to the VFS, we have various accessor-functions for stat, such as
regular stat and lstat.  Same goes for the 'dirstat' functions.  For those
not familiar, those functions are basically "stat a path and convert to a
dir."

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>