rcu: Do not let RCU callbacks block on RCU
The callbacks are run from a ktask.  If the callbacks attempt to block on
RCU with e.g. rcu_barrier() or synchronize_rcu(), we'll deadlock.  The
ktask blocks on RCU callbacks, but the ktask is the only thing that will
run the callbacks.

As far as I can tell, this is also illegal on Linux.  For instance, this
will throw a bunch of errors at runtime on Linux:

static void __cb(struct rcu_head *h)
{
    printk(KERN_INFO "about to sync\n");
    // works only if we comment this out, dies otherwise
    synchronize_rcu();
}

static void foo(void)
{
    struct rcu_head head[1];

    init_rcu_head_on_stack(head);
    call_rcu(head, __cb);
    rcu_barrier();
    destroy_rcu_head_on_stack(head);
}

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>